checking for None type in payload

+++ api/views/stories.py
@@ -12,10 +12,11 @@ STORIES_ID_URL = "/stories/<int:story_id>"
 def get_stories():
     data = request.get_json()
     if 'poi_id' in data:
-        stories = StoryPOI.query.filter(StoryPOI.poi_id == data['poi_id'])
-        if stories.count() == 0:
-            return create_response(status=404, message='No stories found for specified POI')
-        return create_response({'stories': [s.to_dict() for s in stories]})
+        if data['poi_id'] is not None:
+            stories = StoryPOI.query.filter(StoryPOI.poi_id == data['poi_id'])
+            if stories.count() == 0:
+                return create_response(status=404, message='No stories found for specified POI')
+            return create_response({'stories': [s.to_dict() for s in stories]})
 
     stories = Story.query.all()
     return create_response({'stories': [s.to_dict() for s in stories]})
@@ -24,15 +25,19 @@ def get_stories():
 def post_stories():
     data = request.get_json()
     if 'story_name' in data:
-        new_story = row_constructor(Story, data)
+        if data['story_name'] is not None:
+            new_story = row_constructor(Story, data)
+        else:
+            return create_response(status=422, message='The story_name parameter was invalid')
     else:
         return create_response(status=422, message='No story_name parameter was provided')
     db.session.add(new_story)
     db.session.flush()
 
     if 'poi_ids' in data:
-        new_story_pois = [row_constructor(StoryPOI, story_id=new_story.id, poi_id=poi_id) for poi_id in data['poi_ids']]
-        db.session.add_all(new_story_pois)
+        if data['poi_ids'] is not None:
+            new_story_pois = [row_constructor(StoryPOI, story_id=new_story.id, poi_id=poi_id) for poi_id in data['poi_ids']]
+            db.session.add_all(new_story_pois)
     db.session.commit()
     story = Story.query.get(new_story.id)
     return create_response(data={'story': story.to_dict()}, message='Story created')
@@ -40,15 +45,31 @@ def post_stories():
 @app.route(STORIES_ID_URL, methods=['PUT'])
 def put_stories(story_id):
     data = request.get_json()
+    fields = ['story_name', 'poi_ids']
+    missing_params = [field for field in fields if data.get(field) is None]
+
     story = Story.query.get(story_id)
     if 'story_name' in data:
-        story.story_name = data['story_name']
-    if 'poi_ids' in data:
-        StoryPOI.query.filter(StoryPOI.story_id == story_id).delete()
-        if(len(data['poi_ids']) > 0):
-            new_story_pois = [row_constructor(StoryPOI, story_id=story_id, poi_id=poi_id) for poi_id in data['poi_ids']]
-            db.session.add_all(new_story_pois)
+        if data['story_name'] is not None:
+            story.story_name = data['story_name']
             db.session.commit()
+        else:
+            missing_params.append('story_name')
+
+    if 'poi_ids' in data:
+        if data['poi_ids'] is not None:
+            StoryPOI.query.filter(StoryPOI.story_id == story_id).delete()
+            if(len(data['poi_ids']) > 0):
+                new_story_pois = [row_constructor(StoryPOI, story_id=story_id, poi_id=poi_id) for poi_id in data['poi_ids']]
+                db.session.add_all(new_story_pois)
+                db.session.commit()
+        else:
+            missing_params.append('poi_ids')
+
+    if len(missing_params) == 2:
+        message = 'Must provide at least one of the following parameters: story_name or poi_ids'
+        return create_response(data, 422, message)
+
     story = Story.query.get(story_id)
     return create_response(data={'story': story.to_dict()}, message='Story updated')
